make a few minor changes to sqlquery
detectors not active

started to refactor the sql statement for better testing, but then
decided not to do that now.  I tested the heck out of that statement
during development the first time around, so it is probably fine.  I
will revisit the statement later when I have more time to refactor.