notes as I check SQL results
there are maybe two issues at the moment.  The reduced table has cases
where there component is not active from min ts to max endts, but rather
has gaps in time, so therefore should have more than one entry in the
table from mints to maxts.

Second, in sussing that out, I spotted cases with more summed hours than
sum of hours from min to max, implying that there are overlapping
periods of time.  This should nothappen, so I need to write a better
test for the SQL statement in the group_handler code, that proves the
output of the insert statement is what I expect.